+++ api/src/main/java/tpo/uade/api/service/ICartService.java
@@ -6,9 +6,9 @@ import tpo.uade.api.dto.CheckoutDto;
 public interface ICartService {
     void emptyCart(Long userId);
 
-    void addProduct(Long userId, Long productId);
+    void addProduct(Long userId, Integer productId);
 
-    void removeProduct(Long userId, Long productId);
+    void removeProduct(Long userId, Integer productId);
 
     CartDto getCart(Long userId);
